Set minServerVersion for simulcontroller's actions
Most of the actions (the basic ones such as creating posts or visiting
channels) have minServerVersion set to initialVersion, the lowest
version supported by the tool. Some of them, which model features added
later, have the corresponding version set.

Some actions, such as createPost, have version-dependent logic in the
body of the action, so that's handled inside the actions themselves,
while the base action has a lower minServerVersion (initialVersion in
the example of createPost).