Change type of minServerVersion to semver.Version
This, in turn, lets us refactor the way we check for versions. We should
now use semver.MustParse and use semver.Version variables as soon as
possible to reduce parsing errors as much as possible.

This commit also sets an initial version, meaning the lowest version
supported by the controller. As all API requests done by the agents are
based on API v4, any server below v4.0.0 is directly not supported.