Overwrite the server version in tests
Now that the simulcontroller checks the server version when it starts,
we need to make sure that the server the tests talk to return a valid
version. But there's no servers running in the tests! So we just
leverage the ServerVersion field used to overwrite the version itself,
which happens to come in quite handy for this purpose.